requirements.txt
apache-airflow-providers-snowflake>=5.0.0
apache-airflow-providers-amazon>=8.0.0
snowflake-connector-python>=3.0.0
boto3>=1.28.0
requests>=2.31.0
